+++ packages/taro-rn/src/api/Popup/PopupHeader.js
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types'
-import { View, StyleSheet, Text } from 'react-native'
+import { View, StyleSheet, Text, ViewPropTypes } from 'react-native'
 import V from '../variable'
 
 const styles = StyleSheet.create({
@@ -43,7 +43,7 @@ const PopupHeader = ({ style, left = {}, right = {} }) =>
   </View>
 
 PopupHeader.propTypes = {
-  style: View.propTypes.style,
+  style: ViewPropTypes.style,
   left: PropTypes.object,
   right: PropTypes.object
 }