+++ packages/taro-webpack-runner/src/config/base.conf.ts
@@ -2,7 +2,14 @@ import * as path from 'path'
 import * as Chain from 'webpack-chain'
 
 import { getRootPath } from '../util'
-import { getBabelLoader, defaultBabelLoaderOption } from '../util/chain';
+import {
+  getBabelLoader,
+  getUrlLoader,
+  defaultBabelLoaderOption,
+  defaultFontUrlLoaderOption,
+  defaultMediaUrlLoaderOption,
+  defaultImageUrlLoaderOption
+} from '../util/chain';
 
 export default () => {
   const chain = new Chain()
@@ -12,7 +19,6 @@ export default () => {
       rule: {
         jsx: {
           test: /\.jsx?$/,
-          exclude: [/node_modules/],
           use: {
             babelLoader: getBabelLoader([defaultBabelLoaderOption])
           }
@@ -20,31 +26,19 @@ export default () => {
         media: {
           test: /\.(mp4|webm|ogg|mp3|wav|flac|aac)(\?.*)?$/,
           use: {
-            urlLoader: {
-              loader: require.resolve('url-loader')
-            }
+            urlLoader: getUrlLoader([defaultMediaUrlLoaderOption])
           }
         },
         font: {
           test: /\.(woff2?|eot|ttf|otf)(\?.*)?$/,
           use: {
-            urlLoader: {
-              loader: require.resolve('url-loader'),
-              options: {
-                limit: 2000
-              }
-            }
+            urlLoader: getUrlLoader([defaultFontUrlLoaderOption])
           }
         },
         image: {
           test: /\.(png|jpe?g|gif|bpm|svg)(\?.*)?$/,
           use: {
-            urlLoader: {
-              loader: require.resolve('url-loader'),
-              options: {
-                limit: 2000
-              }
-            }
+            urlLoader: getUrlLoader([defaultImageUrlLoaderOption])
           }
         }
       }